requirements.txt
asgiref==3.6.0
asttokens==2.2.1
autopep8==2.0.2
backcall==0.2.0
colorama==0.4.6
crispy-bootstrap4==2022.1
decorator==5.1.1
Django==4.1.7
django-admin-interface==0.24.2
django-ckeditor==6.5.1
django-colorfield==0.8.0
django-crispy-forms==2.0
django-extensions==3.2.1
django-jalali==6.0.1
django-js-asset==2.0.0
executing==1.2.0
extensions==0.4
ipython==8.11.0
jdatetime==4.1.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.38
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.14.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
tzdata==2022.7
wcwidth==0.2.6
